Implemented GOTO, added Macaroni::find_label()
Macaroni::run_tokens() now keeps a mapping of labels to addresses, updating it
whenever a new label is encountered. If a GOTO is encountered, but the label
isn't in the mapping, run_tokens() will search for the label with the new
Macaroni::find_label() and add it to the mapping.

I also did a little bit of cleanup in Macaroni::run_tokens() and execute_op().